vdk-core: add error classification for ImportError
